+++ src/Components/Loading/index.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default () => (
+    <div>
+        Loading Components...
+    </div>
+)
