+++ src/Pages/Home/BuildWith.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export default () => (
+    <div style={{
+        width: "80%",
+        textAlign: "center",
+        padding: "10px",
+        fontSize: "14px",
+    }}
+        className="animate__animated animate__flipInX animate__faster animate__delay-4s">
+        <b>React</b> with <b>Redux</b> and <b>Vite</b> as build tool, hosted on <b>Github Pages</b> with <b>actions</b> on <b>push</b> on <b>main</b>, SSL/TLS by <b>Cloudflare</b> with <b>proxies</b>.
+    </div>
+)
