+++ src/Pages/Home/styles.jsx
@@ -1,30 +1,5 @@
 import styled from "styled-components";
 
-export const ContainerOne = styled.div`
-margin: 0;
-padding: 0;
-height: 100%;
-display: flex;
-flex-flow: column;
-align-items: center;
-justify-content: center;
-@media only screen and (max-width: 812px) {
-    height: 130%;
-}
-`;
-
-
-export const ContainerTwo = styled.div`
-margin: 0;
-margin-bottom: 0.5em;
-padding: 0;
-height: 100vh;
-display: flex;
-flex-flow: column;
-align-items: center;
-justify-content: flex-end;
-`;
-
 export const UlView = styled.ul`
 list-style-type: lower-greek;
 `;
@@ -58,12 +33,4 @@ export const Pre = styled.pre`
 white-space: pre-wrap;
 text-align: center;
 /* word-break: break-all; */
-`;
-
-export const ReduxContainer = styled.div`
-text-align:center;
-border:1px dotted gold;
-border-radius:10px;
-margin-top:10px;
-padding:10px;
 `;
